requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
Django==4.2
django-crispy-forms==2.0
django-session-timeout==0.1.0
idna==3.4
Pillow==9.5.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
